🚚: 공통 컴포넌트 이동
